Added CalendarList and CalendarNotFound classes
Author: Tristen Schuler - tristenschuler98@gmail.com
			  schulert16@students.ecu.edu
Date: 11/13/2019

Added the CalendarList so we can implement mulitple calendars at a time
easily. Added the CalendarNotFound class as an exception for when a
calendar is not found in the list. The old files were deleted and
replaced because I needed to correct my file path on my computer. All
files that were deleted were reuploaded. Also I changed all the
variables in the classes to private, as there is no need for them to be
public when there are get and set methods for them.