Added New Methods to CalendarList
Author: Tristen Schuler - tristenschuler98@gmail.com
Date: 11/14/19

I added new methods to the CalendarList class. replaceCal to replace an
old calendar list with the new one. removeCal to remove a calendar from
the the calendarlist. I also added a new method for addCalendar, where
it takes a calendar to add to the list instead of adding an empty
calendar. May remove the add calendar method where it adds an empty
calendar in the future.