UploadButton in UploadActivity allows user to choose a PDF file
Author: Michael Trinh <trinhm14@students.ecu.edu>
Date: Wed Nov 27 2019
Issue: #9

The upload button in the Syllabye application's UploadActivity.java allows the user to choose a file from the phone's internal storage. I have limited the file selections to .pdf as we are assuming that we have a scenario where we can expect high resolution files. After the user selects the pdf they want, the filepath is shown on the UploadActivity screen so the user knows their file has been selected.