Finished ScanList Class
Author - Tristen Schuler - schulert16@students.ecu.edu |
tristenschuler98@gmail.com

Date - 12/2/2019

Issue - https://github.com/ECU-CSCI-3030/Lucario-Fall-2019/issues/7

Finished ScanList. Added the methods needed for Leo and Jameson to use
it with Tesseract, and for Deric to use it with the UI. Only thing that
needs to be changed in the future is the year and cname variables. Deric
or Michael will be setting it so that they equal what the user gives
them.