Author: Michael Trinh <trinhm14@students.ecu.edu> Date: Sunday Dec 8 2019 Issue: #12
I created a basic camera button that would have been used for the Camera Button feature, issue #12. However, we as a group decided it would be best to scrap this feature as it is not a necessary feature at this moment in time since we have the ability to upload higher resolution PDFs.