Altered TesseractActivity layout and added permissions check functions
Author: Michael Trinh <trinhm14@students.ecu.edu>
Date: Sun Dec 8 2019
Issue: #11 #12

Changed activity_tesseract.xml layout to a scroll view w/ Linearlayouts, altered button/textview layout, added functions to TesseractActivity that will ask the user to confirm permission access to read/write to external storage and to camera. Camera functionality was scrapped but I am implementing a Tesseract OCR that primarily focuses on the camera so I will add some functionality back to this as a possible extra feature.